Move merge_eopatches to eodata_merge (#658)
* move `merge_eopatches` back to `eodata_merge`

* inline `_merge_eopatch_contents` into `merge_eopatches`

* simplify implementation
