+++ core/eolearn/core/__init__.py
@@ -21,7 +21,8 @@ from .core_tasks import (
     SaveTask,
     ZipFeatureTask,
 )
-from .eodata import EOPatch, merge_eopatches
+from .eodata import EOPatch
+from .eodata_merge import merge_eopatches
 from .eoexecution import EOExecutor
 from .eonode import EONode, linearly_connect_tasks
 from .eotask import EOTask
